refactors the error functions to be held in a seperate errors file and moves the checkUserRequest function out of models and into the utils file
